backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.0
neo4j>=5.16.0
google-genai>=1.0.0
google-generativeai>=0.8.0
anthropic>=0.15.0
boto3>=1.34.0
librosa>=0.10.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.26.0
firebase-admin>=6.4.0
python-dotenv>=1.0.0
redis>=5.0.0
tenacity>=8.2.0
PyJWT>=2.8.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
sentry-sdk[fastapi]>=1.40.0
gunicorn>=21.2.0
playwright>=1.40.0





google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
importlib-metadata>=6.0.0; python_version < "3.10"
google-api-core<2.28.0
yt-dlp
